requirements.txt
azure-identity
azure-mgmt-resource
requests